// client/src/types.ts
export type Sex = "Male" | "Female" | "Other";

export interface HouseholdMember {
  id: number;
  name: string;
  sex: Sex;
  heightCm: number;
  weightKg: number;
  bmr: number;
  tdee: number;
  targetCalories: number;
  targetProtein: number;
  avatarColor: string | null;
}
// ...
export interface AddCustomIngredientRequest {
  name: string;
  caloriesPer100g: number;
  proteinPer100g: number;
  carbsPer100g: number;
  fatPer100g: number;
  sugarPer100g: number;
}